requirements.txt
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
streamlit==1.19.0
xgboost==1.7.4